event-radar/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.22.0
python-dotenv>=1.0.0
cloudinary>=1.36.0
pillow>=10.0.0
lxml>=4.9.3
